+++ packages/headless/src/app/logger-middlewares.test.ts
@@ -1,5 +1,6 @@
 import {SchemaValidationError} from '@coveo/bueno';
 import {buildMockSearchAppEngine} from '../test';
+import {serializeSchemaValidationError} from '../utils/validate-payload';
 
 describe('logActionErrorMiddleware', () => {
   it(`when a action has no error parameter
@@ -13,14 +14,35 @@ describe('logActionErrorMiddleware', () => {
     expect(e.logger.error).not.toHaveBeenCalled();
   });
 
-  it(`when a action has an error parameter
+  it(`when a action has a serialized SchemaValidationError
   it should not pass through the middleware`, () => {
     const e = buildMockSearchAppEngine();
     const {dispatch, mockStore: store} = e;
 
-    const action = {type: 'foo', error: new SchemaValidationError('no bueno')};
+    const action = {
+      type: 'foo',
+      error: serializeSchemaValidationError(
+        new SchemaValidationError('no bueno')
+      ),
+    };
     dispatch(action);
     expect(store.getActions()).not.toContain(action);
+  });
+
+  it(`when a action has a serialized Error of any type but SchemaValidationError
+  it should pass through the middleware`, () => {
+    const e = buildMockSearchAppEngine();
+    const {dispatch, mockStore: store} = e;
+
+    const action = {
+      type: 'foo',
+      error: {
+        name: 'TypeError',
+        message: 'You obviously did something wrong',
+      },
+    };
+    dispatch(action);
+    expect(store.getActions()).toContain(action);
   });
 
   it(`when a action has an error parameter
@@ -29,12 +51,16 @@ describe('logActionErrorMiddleware', () => {
     spyOn(e.logger, 'error');
     const {dispatch} = e;
 
-    const error = new SchemaValidationError('no bueno');
-
-    const action = {type: 'foo', error};
+    const error = serializeSchemaValidationError(
+      new SchemaValidationError('no bueno')
+    );
+    const action = {
+      type: 'foo',
+      error,
+    };
     dispatch(action);
     expect(e.logger.error).toHaveBeenCalledWith(
-      error.message,
+      error.stack,
       `Action dispatch error ${action.type}`,
       action
     );