+++ packages/headless/src/app/logger-middlewares.ts
@@ -1,4 +1,4 @@
-import {SchemaValidationError} from '@coveo/bueno';
+import {SerializedError} from '@reduxjs/toolkit';
 import {Logger} from 'pino';
 import {Middleware} from 'redux';
 
@@ -9,11 +9,20 @@ export const logActionErrorMiddleware: (logger: Logger) => Middleware = (
     return next(action);
   }
 
-  const error =
-    typeof action.error === typeof SchemaValidationError
-      ? action.error
-      : action.error.message;
-  logger.error(error, `Action dispatch error ${action.type}`, action);
+  const error: SerializedError = action.error;
+
+  logger.error(
+    error.stack || error.message || error.name || 'Error',
+    `Action dispatch error ${action.type}`,
+    action
+  );
+
+  // Validation errors should prevent further dispatching
+  if (action.error.name === 'SchemaValidationError') {
+    return;
+  }
+
+  return next(action);
 };
 
 export const logActionMiddleware: (logger: Logger) => Middleware = (logger) => (