+++ packages/headless/src/controllers/facets/category-facet/headless-category-facet.ts
@@ -101,16 +101,23 @@ export function buildCategoryFacet(
     /**
      * Selects (deselects) the passed value if unselected (selected).
      * @param selection The category facet value to select or deselect.
+     * @param numberOfValues The number of selection's children to be shown
      */
     toggleSelect: (selection: CategoryFacetValue) =>
-      dispatch(executeToggleCategoryFacetSelect({facetId, selection})),
+      dispatch(
+        executeToggleCategoryFacetSelect({
+          facetId,
+          selection,
+          retrieveCount: options.numberOfValues,
+        })
+      ),
 
     /** Deselects all facet values.*/
     deselectAll: () =>
       dispatch(
         executeDeselectAllCategoryFacetValues({
           facetId,
-          numberOfValues: options.numberOfValues!,
+          numberOfValues: options.numberOfValues,
         })
       ),
 
@@ -161,7 +168,7 @@ export function buildCategoryFacet(
       const request = getRequest();
       const response = getResponse();
 
-      const {parents, values} = partitionIntoParentsAndValues(response);
+      const {parents, values} = partitionIntoParentsAndValues(response?.values);
       const isLoading = engine.state.search.isLoading;
       const hasActiveValues = parents.length !== 0;
       const canShowMoreValues =
@@ -171,6 +178,7 @@ export function buildCategoryFacet(
       const canShowLessValues = values.length > options.numberOfValues;
 
       return {
+        facetId,
         parents,
         values,
         isLoading,